% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skog_functions.R
\name{scale_color_sydskog}
\alias{scale_color_sydskog}
\alias{scale_colour_sydskog}
\alias{scale_colour_sydskog_d}
\alias{scale_color_sydskog_d}
\alias{scale_fill_sydskog}
\alias{scale_fill_sydskog_d}
\title{The sydskog colour scales}
\usage{
scale_color_sydskog(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  discrete = FALSE,
  option = "oak"
)

scale_colour_sydskog(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  discrete = FALSE,
  option = "oak"
)

scale_colour_sydskog_d(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "oak",
  aesthetics = "colour"
)

scale_color_sydskog_d(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "oak",
  aesthetics = "colour"
)

scale_fill_sydskog(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  discrete = FALSE,
  option = "oak"
)

scale_fill_sydskog_d(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "oak",
  aesthetics = "fill"
)
}
\arguments{
\item{...}{parameters to \code{discrete_scale} or \code{scale_fill_gradientn}}

\item{alpha}{pass through parameter to \code{sydskog}}

\item{begin}{The (corrected) hue in from 0 to 1 at which the sydskog colormap begins.}

\item{end}{The (corrected) hue in from 0 to 1 at which the sydskog colormap ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors
are as output by \code{sydskog_pal}. If -1, the order of colors is reversed.}

\item{discrete}{generate a discrete palette? (default: \code{FALSE} - generate continuous palette)}

\item{option}{A character string indicating the colourmap to use.}

\item{aesthetics}{Character string or vector of character strings listing the name(s) of the aesthetic(s) that this scale works with.
This can be useful, for example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the same time, via \code{aesthetics = c("colour", "fill")}.
Parameter inherited from 'ggplot2'.}
}
\description{
Uses the The sydskog color scale.
}
\details{
For \code{discrete == FALSE} (the default) all other arguments are as to
\link[ggplot2]{scale_fill_gradientn} or \link[ggplot2]{scale_color_gradientn}.
Otherwise the function will return a \code{discrete_scale} with the plot-computed
number of colors.
}
\examples{
library(ggplot2)
ggplot(data.frame(x = rnorm(10000), y = rnorm(10000)), aes(x = x, y = y)) +
geom_hex() + coord_fixed() +
scale_fill_sydskog(option="slu") + theme_bw()

ggplot(data.frame(x = rnorm(10000), y = rnorm(10000)), aes(x = x, y = y)) +
geom_hex() + coord_fixed() +
scale_fill_sydskog(option="oak") + theme_bw()


}
