% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skog_functions.R
\name{sydskogMap}
\alias{sydskogMap}
\alias{sydskog}
\alias{sydskog_pal}
\alias{sydskog_palette_names}
\alias{sydskog_colors}
\title{'The sydskog' colour map}
\format{
A data frame containing all the colours used in the palette:
\itemize{
\item V1: Red value
\item V2: Green value
\item V3: Blue value
\item option: Refers to the characters of the world of the sydskog.
}
The sydskog Colour Map.

This function creates a vector of \code{n} equally spaced colors along the
'sydskog colour map'.
}
\usage{
sydskogMap(
  n = 256,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "oak"
)

sydskog(n, alpha = 1, begin = 0, end = 1, direction = 1, option = "oak")

sydskog_pal(alpha = 1, begin = 0, end = 1, direction = 1, option = "oak")

sydskog(n, alpha = 1, begin = 0, end = 1, direction = 1, option = "oak")

sydskog_palette_names()

sydskog_colors()
}
\arguments{
\item{n}{The number of colors (\eqn{\ge 1}) to be in the palette.}

\item{alpha}{The alpha transparency, a number in from 0 to 1, see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{begin}{The (corrected) hue in from 0 to 1 at which the sydskog colormap begins.}

\item{end}{The (corrected) hue in from 0 to 1 at which the sydskog colormap ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors
are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{option}{A character string indicating the colourmap to use. It is not case-sensible}

\item{palettes}{one or more palettes to show}
}
\value{
\code{sydskog} returns a character vector, \code{cv}, of color hex
codes. This can be used either to create a user-defined color palette for
subsequent graphics by \code{palette(cv)}, a \code{col =} specification in
graphics functions or in \code{par}.

\code{sydskogMap} returns a \code{n} lines data frame containing the
red (\code{R}), green (\code{G}), blue (\code{B}) and alpha (\code{alpha})
channels of \code{n} equally spaced colors along the 'The sydskog' colour map.
\code{n = 256} by default.
}
\description{
A dataset containing the colour palettes from the TV show, book series and video games of 'The sydskog'.

A simple function that returns a plot of all available palettes in the \code{thesydskog} - package
}
\examples{



pal <- sydskog(255, option = "oak")
image(volcano, col = pal)

library(ggplot2)

ggplot(data.frame(x = rnorm(10000), y = rnorm(10000)), aes(x = x, y = y)) +
geom_hex() + coord_fixed() +
scale_fill_sydskog(option="oak") + theme_bw()

}
\author{
Lukas Graf \email{graflukas57@gmail.com}

Semi-transparent colors (\eqn{0 < alpha < 1}) are supported only on some
devices: see \code{\link[grDevices]{rgb}}.
}
\keyword{internal}
