% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\docType{data}
\name{sydskog_palettes}
\alias{sydskog_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
sydskog_palettes
}
\description{
Use \code{\link{witcher}} to construct palettes of desired length.
}
\keyword{datasets}
